backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Interview Caller with Twilio and Gemini Integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.116.1",
    "python-decouple>=3.8",
    "twilio>=9.7.2",
    "google-generativeai>=0.8.0",
    "uvicorn>=0.35.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
]
